pyproject.toml
[tool.poetry]
name = "hummer-system-task"
version = "0.1.0"
description = ""
authors = ["Sirojiddin Yakubov <yakubov9791999@gmail.com>"]
readme = "README.md"
packages = [{include = "hummer_system_task"}]

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
certifi = "2023.7.22"
cffi = "1.15.1"
charset-normalizer = "3.2.0"
cryptography = "41.0.3"
defusedxml = "0.7.1"
django = "4.2.4"
django-cors-headers = "4.2.0"
django-filter = "23.2"
django-phonenumber-field = "7.1.0"
django-redis = "5.3.0"
django-templated-mail = "1.1.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
djoser = "2.2.0"
drf-yasg = "1.21.7"
gunicorn = "21.2.0"
idna = "3.4"
inflection = "0.5.1"
infobip-api-python-client = "3.0.3"
install = "1.3.5"
oauthlib = "3.2.2"
packaging = "23.1"
phonenumbers = "8.13.18"
psycopg2-binary = "2.9.7"
pycparser = "2.21"
pyjwt = "2.8.0"
pyotp = "2.9.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python3-openid = "3.2.0"
pytz = "2023.3"
pyyaml = "6.0.1"
redis = "5.0.0"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
six = "1.16.0"
social-auth-app-django = "5.2.0"
social-auth-core = "4.4.2"
sqlparse = "0.4.4"
uritemplate = "4.1.1"
urllib3 = "2.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
